// src/interfaces/i5sing/I5singUser.ts
export interface I5singUser {
    ID: number;
    NN: string;
    I: string;
    B: string;
    E: string;
    M: string;
    CT: number;
    SX: number;
    online: number;
    follow: number;
    Bigv: number;
    Tao: number;
    MC?: number;
    isFollow: number;
    P: number;
    C: number;
    F: boolean;
}